More v2.0 work - Reaching feature parity with 1.0...
Former-commit-id: badd09e22c27c5a2210c53c8d62d716a41793ad7
Former-commit-id: da71ecf66413102ae6c6a75dbea2fa12653f242b

+++ FLAIR 2.0/Maven/flair/src/main/java/com/flair/client/presentation/widgets/GenericWeightSlider.java
@@ -85,7 +85,7 @@ public abstract class GenericWeightSlider extends LocalizedComposite
 	}
 	
 	@UiConstructor
-	public GenericWeightSlider() 
+	public GenericWeightSlider()
 	{
 		super(ClientEndPoint.get().getLocalization());
 		toggle = new MaterialCheckBox();
@@ -97,7 +97,7 @@ public abstract class GenericWeightSlider extends LocalizedComposite
 	}
 	
 	@UiConstructor
-	public GenericWeightSlider(boolean hideSlider) 
+	public GenericWeightSlider(boolean hideSlider)
 	{
 		super(ClientEndPoint.get().getLocalization());
 		toggle = new MaterialCheckBox();
@@ -120,7 +120,7 @@ public abstract class GenericWeightSlider extends LocalizedComposite
 		return toggle.getValue();
 	}
 	
-	public double getWeight() {
+	public int getWeight() {
 		return slider.getValue();
 	}
 	
@@ -128,7 +128,7 @@ public abstract class GenericWeightSlider extends LocalizedComposite
 		return slider.getValue() != SLIDER_MIN_VAL;
 	}
 	
-	public void setEnabled(boolean val, boolean fireEvent) 
+	public void setEnabled(boolean val, boolean fireEvent)
 	{
 		toggle.setValue(val, fireEvent);
 		slider.setEnabled(val);
